refactor: Make validations tidier
