+++ src/lib/utilities/fileDropHandling.js
