+++ src/lib/validation/crossCheckValidation.js
@@ -1,23 +1,21 @@
-import { csvMetadata } from '$lib/stores/generator/CsvStore.ts';
-import { get } from 'svelte/store';
-
 // Validation that runs once both files are dropped and independently validated
 //
 // If validation is succesful => return true
 // If validation is not succesful => return an object with the error message AND an array of errors
 
-export const crossCheckValidation = (parsedCsv, files) => {
+export const crossCheckValidation = (parsedCsv, csvMetadata, imagesFiles) => {
   const attributes = parsedCsv[0];
-  const OK = 1, NOK = -1;
+  const OK = 1,
+    NOK = -1;
   let ipfs_keys = ['image'];
   if (attributes.includes('thumbnail')) ipfs_keys.push('thumbnail');
-  let file_xcheck = files.reduce((a, f) => {
+  let file_xcheck = imagesFiles.reduce((a, f) => {
     a[f.name] = NOK;
     return a;
   }, {});
 
   //---- check for errors in metadata: metadata w/ no file (ghosts)
-  let metadata = get(csvMetadata);
+  let metadata = csvMetadata;
   let errs = [];
   for (const nft_name of Object.keys(metadata)) {
     let nft = metadata[nft_name];
@@ -41,6 +39,7 @@ export const crossCheckValidation = (parsedCsv, files) => {
   if (errs.length === 0) {
     return true;
   } else {
+    console.log(errs);
     return {
       error: 'The images and the CSV do not match',
       errs: errs,
